refactor: ResumePageをリファクタ
