--- src/components/ResumePage/Histories.stories.tsx
@@ -1,14 +1,0 @@
-import { Meta, Story } from '@storybook/react'
-import React from 'react'
-import { Histories, HistoriesProps } from './Histories'
-
-export default {
-  title: 'ResumePage/Histories',
-  component: Histories,
-} as Meta<HistoriesProps>
-
-type Template = Story<HistoriesProps>
-
-const Template: Template = (args) => <Histories {...args} />
-
-export const Default: Template = Template.bind({})
