--- src/components/ResumePage/index.ts
@@ -1,2 +1,0 @@
-export * from './Histories'
-export * from './Skills'
