+++ src/components/page/Resume/index.ts
@@ -1,0 +1,1 @@
+export * from './Resume.page'
